requirements.txt
django
djangorestframework
psycopg2-binary

langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq

pypdf
pgvector

langgraph

wikipedia
ddgs

faiss-cpu
sentence-transformers

groq

streamlit
requests
